% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{abstracts}
\alias{abstracts}
\title{Abstracts of research articles}
\format{
A \code{character} vector of length 100.
}
\usage{
abstracts
}
\description{
A random subset of 100 abstracts of research articles published by the University of Basel.
}
\keyword{datasets}
