% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{auckland_queries}
\alias{auckland_queries}
\title{SDG queries of the University of Auckland}
\format{
A data frame with 16 rows and 4 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\source{
\url{https://www.sdgmapping.auckland.ac.nz/}
}
\usage{
auckland_queries
}
\description{
A dataset containing the SDG queries of \href{https://www.sdgmapping.auckland.ac.nz/}{University of Auckland} (version 1). The queries are available
from \href{https://www.sdgmapping.auckland.ac.nz/}{https://www.sdgmapping.auckland.ac.nz/}. The Auckland queries were developed to build on the
processes developed by the United Nations and the Times Higher Education Ranking in order to create an expanded list of keywords that can be
used to identify SDG-relevant research. There is one query per SDG. There are no queries for SDG-17.
}
\keyword{datasets}
