% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aurora.R
\name{detect_aurora}
\alias{detect_aurora}
\title{Detect Aurora SDG}
\usage{
detect_aurora(corpus, verbose = FALSE)
}
\arguments{
\item{corpus}{object of class \code{tCorpus} containing text.}

\item{verbose}{logical.}
}
\description{
Detect Aurora SDG
}
