% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elsevier.R
\name{detect_elsevier}
\alias{detect_elsevier}
\title{Detect Elsevier SDG}
\usage{
detect_elsevier(corpus, verbose = FALSE)
}
\arguments{
\item{corpus}{object of class \code{tCorpus} containing text.}

\item{verbose}{logical.}
}
\description{
Detect Elsevier SDG
}
