% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology.R
\name{detect_ontology}
\alias{detect_ontology}
\title{Detect ontology SDG}
\usage{
detect_ontology(corpus, verbose = FALSE)
}
\arguments{
\item{corpus}{object of class \code{tCorpus} containing text.}

\item{verbose}{logical.}
}
\description{
Detect ontology SDG
}
