% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect_sdg}
\alias{detect_sdg}
\title{Detect SDGs}
\usage{
detect_sdg(
  ...,
  system = c("aurora", "elsevier", "siris", "ontology"),
  out = c("features", "docs"),
  verbose = TRUE
)
}
\arguments{
\item{verbose}{logical.}

\item{corpus}{object of class \code{tCorpus} containing text.}
}
\description{
Detect SDGs
}
