% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{detect_sdg_ensemble}
\alias{detect_sdg_ensemble}
\title{Detect SDGs in text using ensemble model}
\usage{
detect_sdg_ensemble(text, sdgs = 1:17, synthetic = "equal", verbose = TRUE)
}
\arguments{
\item{text}{\code{character} vector or object of class \code{tCorpus} containing text in which SDGs shall be detected.}

\item{sdgs}{\code{numeric} vector with integers between 1 and 17 specifying the sdgs to identify in \code{text}. Defaults to \code{1:17}.}

\item{synthetic}{\code{character} specifying the ensemble version in terms of the amount of synthetic data used in training (relative to the amount of expert-labeled data). Can be one of \code{"none"}, \code{"third"}, \code{"equal"}, and \code{"tripple"}. The default is \code{"equal"}.}

\item{verbose}{\code{logical} specifying whether messages on the function's progress should be printed.}
}
\value{
The function returns a tibble containing the SDG hits found in the vector of documents. Depending on the value of \code{output} the tibble will contain all or some of the following columns:
\describe{
 \item{document}{Index of the element in \code{text} where match was found. Formatted as a factor with the number of levels matching the original number of documents.}
 \item{sdg}{Label of the SDG found in document.}
 \item{system}{The name of the system that produced the match. In this case \code{"Ensemble"}.}
 \item{hit}{Index of hit for the Ensemble model.}
}
}
\description{
\code{detect_sdg_ensemble} identifies SDGs in text using an ensemble model approach considering multiple existing SDG query systems and text length.
}
\details{
\code{detect_sdg_ensemble} implements a ensemble model to detect SDGs in text. The ensemble model combines the six systems implemented by \code{\link{detect_sdg}} and text length in a random forest architecture. The ensemble model has been trained on three data sets with SDG labels assigned by experts and a matching number of synthetic texts generated by random sampling from a word frequency list. The user has the choice of multiple versions of the ensemble model that have been trained on different amounts of synthetic texts to adjust the sensitivity and specificity of the model. Increasing the amount of of synthetic data makes the ensemble more conservative, leading to increased sensitivity and decreased specificity.

By default, \code{detect_sdg_ensemble} implements an ensemble model that has been trained on an equal amount of expert-labeled and synthetic data, providing a reasonable balance between sensitivity and specificity. For details, see article by Wulff et al. (2023).
}
\examples{
\donttest{
# run sdg detection
hits <- detect_sdg_ensemble(projects)

# run sdg detection for sdg 3 only
hits <- detect_sdg_ensemble(projects, sdgs = 3)
}
}
\references{
Wulff, D. U., Meier, D., & Mata, R. (2023). Using novel data and ensemble models to improve automated SDG-labeling. arXiv
}
