% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siris.R
\name{detect_siris}
\alias{detect_siris}
\title{Detect Siris SDG}
\usage{
detect_siris(corpus, verbose = FALSE)
}
\arguments{
\item{corpus}{object of class \code{tCorpus} containing text.}

\item{verbose}{logical.}
}
\description{
Detect Siris SDG
}
