% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{make_corpus}
\alias{make_corpus}
\title{Make Corpus}
\usage{
make_corpus(txt, ...)
}
\arguments{
\item{txt}{character vector}

\item{...}{further arguments for corpustools::search_features}
}
\description{
Makes Corpus
}
