% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ontology_queries}
\alias{ontology_queries}
\title{SDG keyword ontology of Bautista-Puig and Mauleón}
\format{
A data frame with 4,122 rows and 5 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{keyword}{SDG keyword used in query}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\usage{
ontology_queries
}
\description{
A dataset containing the SDG queries based on the keyword ontology of Bautista-Puig and Mauleón. The queries are available from \href{https://figshare.com/articles/dataset/SDG_ontology/11106113/1}{figshare.com}.
}
\details{
Bautista-Puig, N.; Mauleón E. (2019). Unveiling the path towards sustainability: is there a research interest on sustainable goals? In the 17th International Conference on Scientometrics & Informetrics (ISSI 2019), Rome (Italy), Volume II, ISBN 978-88-3381-118-5, p.2770-2771.
}
\keyword{datasets}
