% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{projects}
\alias{projects}
\title{Descriptions of research projects}
\format{
A \code{character} vector of length 500.
}
\source{
\url{https://data.snf.ch/about/glossary}
}
\usage{
projects
}
\description{
500 project descriptions of University of Basel research projects that were funded by the Swiss National Science Foundation. The project descriptions were drawn randomly from University of Basel projects listed in the the public \href{https://data.snf.ch}{data.snf.ch project data base}.
}
\keyword{datasets}
