% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdsn_queries}
\alias{sdsn_queries}
\title{SDG keywords by SDSN}
\format{
A data frame with 847 rows and 5 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{keyword}{SDG keyword used in query}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\usage{
sdsn_queries
}
\description{
DESCRIPTION
}
\details{
DESCRIPTION 2
}
\keyword{datasets}
