% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{search_corpus}
\alias{search_corpus}
\title{Search Corpus}
\usage{
search_corpus(corpus, queries, ...)
}
\arguments{
\item{corpus}{object of class \code{tCorpus} containing text.}

\item{queries}{character vector if queries}

\item{...}{further arguments for corpustools::search_features}
}
\description{
Search Corpus
}
