% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aurora_queries}
\alias{aurora_queries}
\title{SDG queries of the Aurora Universities Network}
\format{
A data frame with 378 rows and 5 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{sdg_title}{Title of the SDG}
  \item{sdg_description}{Description of the SDG}
  \item{query_id}{Index of the query}
  \item{query}{Original SDG query}
}
}
\usage{
aurora_queries
}
\description{
A dataset containing the SDG queries version 5.0 of the \href{https://aurora-network.global/activity/sustainability/}{Aurora Universities Network}. See the corresponding \href{https://github.com/Aurora-Network-Global/sdg-queries}{GitHub repository}. For the actual implementation of the queries see \code{aurora_simple}, \code{aurora_and}, \code{aurora_w}, and the queries hard-coded on \code{detect_aurora}.
}
\keyword{datasets}
