% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{elsevier_queries}
\alias{elsevier_queries}
\title{SDG queries of Elsevier}
\format{
A data frame with 378 rows and 5 columns
\describe{
  \item{sdg}{Label of the SDG}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\usage{
elsevier_queries
}
\description{
A dataset containing the SDG queries of \href{https://www.elsevier.com/connect/sdg-report}{Elsevier}. The queries are available from \href{https://data.mendeley.com/datasets/87txkw7khs/1}{data.mendeley.com}.
}
\keyword{datasets}
