% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{elsevier_queries}
\alias{elsevier_queries}
\title{SDG queries of Elsevier}
\format{
A data frame with 16 rows and 4 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\source{
\url{https://data.mendeley.com/datasets/87txkw7khs/1}
}
\usage{
elsevier_queries
}
\description{
A dataset containing the SDG queries of \href{https://www.elsevier.com/connect/sdg-report}{Elsevier} (version 1). The queries are available from \href{https://data.mendeley.com/datasets/87txkw7khs/1}{data.mendeley.com}. The Elsevier queries were developed to maximize SDG hits on the Scopus database. A detailed description of how each SDG query was developed can be found \href{https://elsevier.digitalcommonsdata.com/datasets/87txkw7khs/1}{here}. There is one query per SDG. There are no queries for SDG-17.
}
\keyword{datasets}
