% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_sdg}
\alias{plot_sdg}
\title{Plot SDGs}
\usage{
plot_sdg(
  hits,
  show_sdg = 1:17,
  show_system = c("aurora", "elsevier", "siris"),
  color = "unibas",
  sdg_titles = FALSE,
  remove_duplicates = TRUE
)
}
\arguments{
\item{hits}{object of class \code{tCorpus} containing text.}
}
\description{
Plot SDGs
}
