% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sdsn_queries}
\alias{sdsn_queries}
\title{SDG keywords by SDSN}
\format{
A data frame with 847 rows and 5 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{keyword}{SDG keyword used in query}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\source{
\url{https://ap-unsdsn.org/regional-initiatives/universities-sdgs/}
}
\usage{
sdsn_queries
}
\description{
A dataset containing SDG-specific keywords compiled from several universities from the Sustainable Development Solutions Network (SDSN) Australia, New Zealand & Pacific Network. The authors used UN documents, Google searches and personal communications as sources for the keywords. All SDGs (1-17) are covered.
}
\keyword{datasets}
