% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{siris_queries}
\alias{siris_queries}
\title{SDG queries of SIRIS Academic}
\format{
A data frame with 3,445 rows and 6 columns
\describe{
  \item{system}{Name of system}
  \item{sdg}{Label of the SDG}
  \item{keyword}{Primary SDG query element}
  \item{extra}{Secodary SDG query element}
  \item{query_id}{Index of the query}
  \item{query}{SDG query}
}
}
\usage{
siris_queries
}
\description{
A dataset containing the SDG queries of \href{http://www.sirislab.com/lab/sdg-research-mapping/}{SIRIS Academic}. The queries are available from\href{https://zenodo.org/record/3567769#.YRY9xdMzY8N}{Zenodo.org}. For the actual implementation of the queries see \code{aurora_simple}, \code{aurora_and}, \code{aurora_w}, and the queries hard-coded on \code{detect_aurora}.
}
\keyword{datasets}
